Write views that actually do something
Each view is responsible for doing one of two things: returning an
HttpResponse object containing the content for the requested page, or
raising an exception such as Http404.  The rest is up to you.

Your view can read records from a database, or not.  It can use a
template system such as Django's --- or a third-party Python template
system --- or not.  It can generate a PDF file, output XML, create a ZIP
file on the fly, anything you want, using whatever Python libraries you
want.

All Django wants is that HttpResponse, or an exception.

In this commit, we rewrote the index() view using Django's own database
API, to display the latest 5 poll questions in the system, separated by
newlines, according to publication date.