Replace the template and the loader with the shortcut render() function
It's a very common idiom to load a template, fill a context and return
an HttpResponse object with the result of the rendered template.  Django
provides a shortcut render() function to do this.

The render() function takes the request object as its first argument, a
template name as its second argument and a dictionary as its optional
third argument.  It returns an HttpResponse object of the given template
rendered with the given context.