Use get_object_or_404() instead of manually raising exceptions
It's a very common idiom to use get() and raise Http404 if the object
doesn't exist.  Django provides a shortcut function get_object_or_404().

The get_object_or_404() function takes a Django model as its first
argument and an arbitrary number of keyword arguments, which it passes
to the get() function of the model's manager.  It raises Http404 if the
object doesn't exist.

Why do we use a helper function get_object_or_404() instead of
automatically catching the ObjectDoesNotExist exception at a
higher-level, or having the model API raise Http404 instead of
ObjectDoesNotExist?

Because that would couple the model layer to the view layer.  One of the
foremost design goals of Django is to maintain loose coupling.  Some
controlled coupling is introduced in the django.shortcuts module.

There's also a get_list_or_404() function, which works just as
get_object_or_404() --- except using filter() instead of get().  It
raises Http404 if the list is empty.