Use the template system
The template system uses dot-lookup syntax to access variale attributes.
In the example of {{question.question_text}}, first Django does a
dictionary lookup on the object question.  Failing that, it tries an
attribute lookup --- which works, in this case.  If attribute lookup had
failed, it would've tried a list-index lookup.

Method-calling happens in the {% for %} loop: question.choice_set.all is
interpreted as the Python code question.choice_set.all(), which returns
an iterable of Choice objects and is suitable for use in the {% for %}
tag.

See the template guide for more about templates.