Add namespace so Django can differentiate URL names between apps
The tutorial project has just one app, polls.  In real Django projects,
there might be five, ten, twenty apps or more.  How does Django
differentiate the URL names between them?  For example, the polls app
has a detail view, and so might an app on the same project that is for a
blog.  How does one make it so that Django knows which app view to
create for a URL when using the {% url %} template tag?

The answer is to add namespacs to your URLconf.  In the polls/urls.py
file, go ahead and add an app_name to set the application namespace:

    app_name = 'polls'

Now change your polls/templates/polls/index.html to:

    <li>
      <a href="{% url 'polls:detail' question.id %}">
        {{question.question_text}}
      </a>
    </li>