Write a simple form
The "polls/detail.html" template displays a radio button for each
question choice.  The value of each radio button is the associated
question's ID.  The name of each radio button is "choice".  That means,
when somebody selectes one of the radio buttons and submits the form,
it'll send the POST data choice=xx where xx is the ID of the selected
choice.  This is the concept of HTML forms.

We set the form's action to {% url 'polls:vote' question.id %}, and we
set method="post".  Using method="post" (as opposed to method="get") is
very important, because the act of submitting this form will alter data
server-side.  Whenever you create a form that alters data server-side,
use method="post".  This tip isn't specific to Django; it's just good
Web development practice.

forloop.counter indicates how many times the for tag has gone through
its loop.

Since we're creating a POST form (which can have the effect of modifying
data), we need to worry about Cross Site Request Forgeries.  Thankfully,
you don't have to worry too hard, because Django comes with a very
easy-to-use system for protecting against it.  In short, all POST forms
that are targeted at internal URLs should use the {% csrf_token %}
template tag.

request.POST is a dictionary-like object that lets you access submitted
data by key name.  In this case, request.POST['choice'] returns the ID
of the selected choice, as a string.  request.POST values are always
strings.

Not that Django also provides request.GET for accessing GET data in the
same way --- but we're explicitly using request.POST in our code, to
ensure that data is only altered via a POST call.

request.POST['choice'] will raise KeyError if choice wasn't provided in
POST data.  The above code checks for KeyError and redisplays the
question form with an error message if choice isn't given.

After incrementing the choice count, the code returns an
HttpResponseRedirect rather than a normal HttpResponse.
HttpResponseRedirect takes a single argument: the URL to which the user
will be redirected (see the following point for how we construct the URL
in this case).

As the Python comment above points out, you should always return an
HttpResponseRedirect after successfully dealing with POST data

We are using the reverse() function in the HttpResponseRedirect
constructor in this example.  This function helps avoid having to
hardcode a URL in the view function.  It is given the name of the view
that we want to pass control to and the variable portion of the URL
pattern that points to that view.  In this case, using the URLconf we
set up in Tutorial 3, this reverse() call will return a string like

    /polls/3/results/

where the 3 is the value of question.id.  This redirected URL will then
call the 'results' view to display the final page.

NOTE: the code for our vote() view does have a small problem.  It first
gets the selected_choice object from the database, then computes the new
value of votes, and then saves it back to the database.  If two users of
your website try to vote at EXACTLY THE SAME TIME, this might go wrong:
the same value let's say 42, will be retrieved for votes.  Then, for
both users the new value of 43 is computed and saved, but 44 would be
the expected value.

This is called RACE CONDITION.  If you are interested, you can read
"Avoiding race conditions using F()" to learn how you can solve this
issue.