Write our first test and then fix a bug found by the test
There's a little bug in the polls application for us to fix right away:
the Question.was_published_recently() method returns True if the
question was published within the last day (which is correct) but also
if the question's pub_date field is in the future (which certainly
isn't).

An conventional place for an application's tests is in the application's
tests.py file; the testing system will automatically find tests in any
file whose name begins with test.

What happens is this:

  - ./manage.py test polls looked for tests in the polls application
  - it found a subclass of the django.test.TestCase class
  - it created a special database for the purpose of testing
  - it looked for test methods --- ones whose names begin with test
  - in test_was_published_recently_with_future_question it created a
    Question instance whose pub_date field is 30 days in the future
  - ...and using the assertIs() method, it discovered that its
    was_published_recently() returns True, though we wanted it to return
    False.